+++ frontend/src/views/Seedlot/SeedlotRegFormClassA/constants.ts
