+++ frontend/src/views/Seedlot/SeedlotRegFormClassA/definitions.ts
