+++ frontend/src/views/Seedlot/SeedlotRegFormClassA/utils.ts
